- Work on getting FLV video working. It should be simple, but it's not   (rant suppressed). I can get a swf that plays video working... - Decided to change location coordinates to simply be floating point numbers   between 0 and 1. All rooms will have these coordinates, and the size of the   room will be an attribute of the scene. In this way, furniture positions   do not need to be altered if a user upgrades or downgrades a room.

+++ src/java/com/threerings/msoy/world/data/MsoyLocation.java
@@ -31,12 +31,12 @@ public class MsoyLocation
     /**
      * Constructs a fully-specified Location.
      */
-    public MsoyLocation (float x, float y, float z, short orient)
+    public MsoyLocation (double x, double y, double z, int orient)
     {
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.orient = orient;
+        this.x = (float) x;
+        this.y = (float) y;
+        this.z = (float) z;
+        this.orient = (short) orient;
     }
 
     /**
